Fix for the error: RuntimeError: Expected object of scalar type Double but got scalar type Float for argument #3 'mat2' in call to _th_addmm_out. This error occurs in the notebook: lesson3-head-pose.ipynb.

+++ fastai/vision/image.py
@@ -252,7 +252,7 @@ class ImagePoints(Image):
     "Support applying transforms to a `flow` of points."
     def __init__(self, flow:FlowField, scale:bool=True, y_first:bool=True):
         if scale: flow = scale_flow(flow)
-        if y_first: flow.flow = flow.flow.flip(1)
+        if y_first: flow.flow = flow.flow.flip(1).float()
         self._flow = flow
         self._affine_mat = None
         self.flow_func = []
